+++ packages/vscode-framework/test/cli/buildConfiguration.test.ts
@@ -1,5 +1,0 @@
-/// <reference types="jest" />
-
-import { runConfigurationGenerator } from '../../src/cli/configurationFromType'
-
-test.todo('Generates config from type')